Achieved modularisation without breaking the program
