+++ app/src/Eval.hs-boot
@@ -1,0 +1,7 @@
+module Eval where
+
+import Lisp.Val
+import Util.Flow
+import {-# SOURCE #-} Env
+
+eval :: Env -> LispVal -> IOThrowsError LispVal
