+++ app/src/Repl.hs
@@ -4,22 +4,19 @@ import Control.Monad
 import System.IO
 
 import Lisp.Val
-import Parser
-import Eval
 import Lisp.Error
+
+import Util.Flow
+
+import Parser
 import Env
+import Eval
 
 flushStr :: String -> IO ()
 flushStr str = putStr str >> hFlush stdout
 
 readPrompt :: String -> IO String
 readPrompt prompt = flushStr prompt >> getLine
-
-evalString :: Env -> String -> IO String
-evalString env expr = runIOThrows (liftThrows (readExpr expr) >>= eval env . show)
-
-evalAndPrint :: Env -> String -> IO ()
-evalAndPrint env expr = evalString env expr >>= putStrLn
 
 until_ :: Monad m => (a -> Bool) -> m a -> (a -> m ()) -> m ()
 until_ pred prompt action = do
