+++ app/src/Util/Flow.hs
@@ -1,0 +1,23 @@
+module Util.Flow where
+
+import Control.Monad.Except
+
+import Lisp.Error
+
+type ThrowsError = Either LispError
+
+trapError :: (MonadError e m, Show e) => m String -> m String
+trapError action = catchError action (return . show)
+
+extractValue :: ThrowsError a -> a
+extractValue (Right val) = val
+
+type IOThrowsError = ExceptT LispError IO
+
+-- "Recontextualises" errors in IO environments 
+liftThrows :: ThrowsError a -> IOThrowsError a
+liftThrows (Left err) = throwError err
+liftThrows (Right val) = return val
+
+runIOThrows :: IOThrowsError String -> IO String
+runIOThrows action = extractValue <$> runExceptT (trapError action)
