+++ app/src/Util/Unpacker.hs
@@ -5,6 +5,7 @@ import Data.Char
 
 import Lisp.Val
 import Lisp.Error
+import Util.Flow
 
 data Unpacker = forall a. Eq a => AnyUnpacker (LispVal -> ThrowsError a)
